datagen/thelook-ecomm/requirements.txt
faker
psycopg2-binary
SQLAlchemy
kafka-python
